refactor : OutputView 클래스의 결과 출력 메소드를 Double형에 맞게 수정

+++ src/main/java/com/woowacourse/calculator/view/OutputView.java
@@ -1,7 +1,7 @@
 package com.woowacourse.calculator.view;
 
 public class OutputView {
-    public void printCalculateResult(final int calculateResult) {
+    public void printCalculateResult(final Double calculateResult) {
         System.out.println("계산 결과: " + calculateResult);
     }
 }
